Create render thread, cleaner code structure
