+++ finch/render.py
@@ -1,0 +1,69 @@
+import cv2
+import time
+
+from finch.shared_state import State
+
+WINDOW_NAME = "drarwing_continuous"
+DIFF_WINDOW_NAME = "drarwing_continuous_diff"
+MIN_FRAME_DURATION = 1/60
+
+
+def _window_exists(window_name: str):
+    try:
+        return cv2.getWindowProperty(window_name, 0) >= 0
+    except:
+        return False
+
+
+def render_thread(shared_state: State, fullscreen: bool = True, show_diff: bool = False, debug: bool = False):
+    if fullscreen:
+        cv2.namedWindow(WINDOW_NAME, cv2.WINDOW_NORMAL)
+        cv2.setWindowProperty(WINDOW_NAME, cv2.WND_PROP_FULLSCREEN, cv2.WINDOW_FULLSCREEN)
+    else:
+        cv2.namedWindow(WINDOW_NAME)
+
+    if show_diff:
+        cv2.namedWindow(DIFF_WINDOW_NAME)
+
+    last_frame = time.time()
+    last_frame_duration = -1.0
+    while (
+        not shared_state.flag_stop
+        and _window_exists(WINDOW_NAME)
+        and (not show_diff or _window_exists(DIFF_WINDOW_NAME))
+    ):
+        if shared_state.image_available:
+            if debug:
+                img = shared_state.specimen.cached_image.copy()
+                img = cv2.putText(
+                    img,
+                    f"{shared_state.img_path}-{shared_state.brush.name} "
+                    f"{shared_state.update_time_microseconds} ms, {shared_state.score} score "
+                    f'{round(1 / last_frame_duration) if last_frame_duration > 0 else "inf"} fps',
+                    (10, 20),
+                    cv2.FONT_HERSHEY_SIMPLEX,
+                    0.5,
+                    (0, 0, 255),
+                    1,
+                    cv2.LINE_AA
+                )
+                cv2.imshow(WINDOW_NAME, img)
+            else:
+                cv2.imshow(WINDOW_NAME, shared_state.specimen.cached_image)
+
+            if show_diff:
+                cv2.imshow(DIFF_WINDOW_NAME, shared_state.specimen.diff_image)
+
+            key = cv2.waitKey(1)
+            if key > -1:
+                match key:
+                    case 110: shared_state.flag_next_image = True  # "n"
+                    case _: break
+
+        now = time.time()
+        last_frame_duration = now - last_frame
+        last_frame = now
+        if last_frame_duration < MIN_FRAME_DURATION:
+            time.sleep(MIN_FRAME_DURATION - last_frame_duration)
+
+    shared_state.flag_stop = True
